enhancement(addons): added upload addons dialog to Addons section
