+++ src/pages/SettingsPage/Bundles/Bundles.jsx
@@ -2,7 +2,7 @@ import { useState, useMemo, useEffect } from 'react'
 import BundleList from './BundleList'
 import BundleDetail from './BundleDetail'
 
-import { Button, InputSwitch, Section, Dialog } from '@ynput/ayon-react-components'
+import { Button, InputSwitch, Section } from '@ynput/ayon-react-components'
 import * as Styled from './Bundles.styled'
 import { useGetBundleListQuery } from '/src/services/bundles/getBundles'
 import {
@@ -15,7 +15,7 @@ import { useGetInstallerListQuery } from '/src/services/installers'
 import { useGetAddonListQuery } from '../../../services/addons/getAddons'
 import { upperFirst } from 'lodash'
 import { toast } from 'react-toastify'
-import AddonUpload from '../AddonInstall/AddonUpload'
+import AddonDialog from '../../../components/AddonDialog/AddonDialog'
 import { useGetAddonSettingsQuery } from '/src/services/addonSettings'
 import getLatestSemver from './getLatestSemver'
 import { ayonApi } from '/src/services/ayon'
@@ -33,9 +33,6 @@ const Bundles = () => {
   const dispatch = useDispatch()
   // addon install dialog
   const [uploadOpen, setUploadOpen] = useState(false)
-
-  // keep track is an addon was installed
-  const [restartRequired, setRestartRequired] = useState(false)
 
   // table selection
   const [selectedBundles, setSelectedBundles] = useState([])
@@ -295,12 +292,6 @@ const Bundles = () => {
       },
     })
 
-  const handleAddonInstallFinish = () => {
-    setUploadOpen(false)
-    if (restartRequired) {
-      setRestartRequired(false)
-    }
-  }
 
   let uploadHeader = ''
   switch (uploadOpen) {
@@ -361,21 +352,7 @@ const Bundles = () => {
         shortcuts={shortcuts}
         deps={[selectedBundles, newBundleOpen, prodBundle, stageBundle]}
       />
-      <Dialog
-        isOpen={uploadOpen}
-        style={{ width: 400, height: 400, overflow: 'hidden' }}
-        header={uploadHeader}
-        onClose={handleAddonInstallFinish}
-        size="md"
-      >
-        {uploadOpen && (
-          <AddonUpload
-            onClose={handleAddonInstallFinish}
-            type={uploadOpen}
-            onInstall={(t) => t === 'addon' && setRestartRequired(true)}
-          />
-        )}
-      </Dialog>
+      <AddonDialog uploadOpen={uploadOpen} setUploadOpen={setUploadOpen} uploadHeader={uploadHeader} />
       <main style={{ overflow: 'hidden' }}>
         <Splitter style={{ width: '100%' }} stateStorage="local" stateKey="bundles-splitter">
           <SplitterPanel style={{ minWidth: 200, width: 400, maxWidth: 800, zIndex: 10 }} size={30}>
